Use unix command 'strings' instead of ruby gem 'roo'
Since we only need to find emails in excel sheets we don't need any fancy lib to read worksheets and so on. This also solves problem with reading big excel files that consumed a lot of memory in the ruby process.

By using the unix command 'strings' we collect emails by just read all strings in a "binary" file.

Later version of excel is a zip-file with xml-files. In those cases we unzip the file and then parses the files with 'strings'.

 - Fixes bug with not being able to parse 95-excel files
 - When comparing expected emails in spec sort them before check